chore(audio): move Harmonizer to audio package
